draft option question popup
